pyproject.toml
[project]
name = "agentic-workflow"
version = "0.1.0"
description = "Demonstration of an agentic workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "mistralai>=1.9.7",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.10",
    "tavily-python>=0.7.11",
]
